Extract audio settings popover
